@@ -12,7 +12,7 @@ module Testpoint
     config.app_generators.scaffold_controller :responders_controller
 
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 6.0
+    config.load_defaults 7.0
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
@@ -25,5 +25,6 @@ module Testpoint
     end
 
     config.time_zone = "Beijing"
+    config.active_record.use_yaml_unsafe_load = true
   end
 end
