@@ -1,2 +1,2 @@
 # Set default locale to something other than :en
-I18n.default_locale = :"zh-CN"
+I18n.default_locale = :'zh-CN'
