@@ -13,7 +13,7 @@ class MigrateToStorage < ActiveRecord::Migration[5.2]
 
   def data_migration(attachment)
     return if ActiveStorage::Attachment.where(record_id: attachment.id, record_type: "Attachment").exists?
-    path = Rails.root.join('public', 'uploads', 'attachment', 'file', attachment.id.to_s, attachment.file)
+    path = Rails.root.join('public/uploads/attachment/file', attachment.id.to_s, attachment.file)
     content_type = Mime::Type.lookup_by_extension(path.extname.delete('.')).to_s
     file = Rack::Test::UploadedFile.new(path, content_type)
     blob = ActiveStorage::Blob.create_after_upload!(io: file, filename: file.original_filename, content_type: file.content_type)
